backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pillow>=10.3.0
numpy>=1.26.4
sqlalchemy>=2.0.30
python-dotenv>=1.0.1
aiofiles>=23.2.1
python-multipart>=0.0.9
scipy>=1.11.0
supabase>=2.0.0
